requirements.txt
streamlit
openai 
azure-search-documents 
python-dotenv